adds (optional) gravity & absorption band features; adds explosive CMT feature for axisymmetric simulations; updates tests

+++ .github/scripts/run_tests.sh
@@ -8,6 +8,8 @@ if [ -f $HOME/.tmprc ]; then source $HOME/.tmprc; fi
 
 WORKDIR=`pwd`
 dir=${TESTDIR}
+TESTID=${TESTID:-}
+TESTCOV=${TESTCOV:-}
 
 # info
 echo "work directory: $WORKDIR"
@@ -22,12 +24,14 @@ echo
 
 # bash function for checking seismogram output with reference solutions
 my_test(){
+  echo "######################################################################################################################"
   echo "testing seismograms:"
   ln -s $WORKDIR/utils/compare_seismogram_correlations.py
   ./compare_seismogram_correlations.py REF_SEIS/ OUTPUT_FILES/
   if [[ $? -ne 0 ]]; then exit 1; fi
   ./compare_seismogram_correlations.py REF_SEIS/ OUTPUT_FILES/ | grep min/max | cut -d \| -f 3 | awk '{print "correlation:",$1; if ($1 < 0.999 ){print $1,"failed"; exit 1;}else{ print $1,"good"; exit 0;}}'
   if [[ $? -ne 0 ]]; then exit 1; fi
+  echo "######################################################################################################################"
 }
 
 # test example
@@ -50,20 +54,36 @@ if [ "${TESTDIR}" == "EXAMPLES/applications/axisymmetric_examples/axisymmetric_c
   sed -i "s:^NPROC .*:NPROC    = 2:" DATA/Par_file
   sed -i "s:^NSTEP .*:NSTEP    = 1400:" DATA/Par_file
 fi
+if [ "${TESTDIR}" == "EXAMPLES/reproducible_study/Komatitsch2000_fluid_solid/fluid_solid_external_mesh" ]; then
+  sed -i "s:^NPROC .*:NPROC    = 2:" DATA/Par_file
+fi
+if [ "${TESTDIR}" == "EXAMPLES/reproducible_study/Morency2008_poroelastic_semi_infinite_homogeneous" ]; then
+  sed -i "s:^NSTEP .*:NSTEP    = 2000:" DATA/Par_file
+fi
+if [ "${TESTDIR}" == "EXAMPLES/applications/Rayleigh_wave_no_crack" ]; then
+  sed -i "s:^NPROC .*:NPROC    = 4:" DATA/Par_file
+fi
+if [ "${TESTDIR}" == "EXAMPLES/applications/Rayleigh_wave_with_crack" ]; then
+  sed -i "s:^NPROC .*:NPROC    = 4:" DATA/Par_file
+fi
 
 
 # debug
 if [ "${DEBUG}" == "true" ]; then
   # limit for debugging
-  sed -i "s:^NSTEP .*:NSTEP = 5:" DATA/Par_file
+  sed -i "s:^NSTEP .*:NSTEP    = 5:" DATA/Par_file
+fi
+
+# coverage runs use short steps
+if [ "$TESTCOV" == "true" ]; then
+  sed -i "s:^NSTEP .*:NSTEP    = 10:" DATA/Par_file
 fi
 
 # selects kernel script for kernel benchmark examples
-do_kernel_script=0
-
-if [ "${TESTDIR}" == "EXAMPLES/benchmarks/BENCHMARK_CLAERBOUT_ADJOINT/ACOUSTIC" ]; then do_kernel_script=1; fi
-if [ "${TESTDIR}" == "EXAMPLES/benchmarks/BENCHMARK_CLAERBOUT_ADJOINT/ELASTIC" ]; then do_kernel_script=1; fi
-if [ "${TESTDIR}" == "EXAMPLES/benchmarks/BENCHMARK_CLAERBOUT_ADJOINT/ACOUSTIC_ELASTIC" ]; then do_kernel_script=1; fi
+RUN_KERNEL=${RUN_KERNEL:-}
+if [ "${TESTDIR}" == "EXAMPLES/benchmarks/BENCHMARK_CLAERBOUT_ADJOINT/ACOUSTIC" ]; then RUN_KERNEL=true; fi
+if [ "${TESTDIR}" == "EXAMPLES/benchmarks/BENCHMARK_CLAERBOUT_ADJOINT/ELASTIC" ]; then RUN_KERNEL=true; fi
+if [ "${TESTDIR}" == "EXAMPLES/benchmarks/BENCHMARK_CLAERBOUT_ADJOINT/ACOUSTIC_ELASTIC" ]; then RUN_KERNEL=true; fi
 
 # setup elastic kernel for SH simulations
 if [ "${TESTDIR}" == "EXAMPLES/benchmarks/BENCHMARK_CLAERBOUT_ADJOINT/ELASTIC" ] && [ "${TESTCASE}" == "SH" ]; then
@@ -71,15 +91,19 @@ if [ "${TESTDIR}" == "EXAMPLES/benchmarks/BENCHMARK_CLAERBOUT_ADJOINT/ELASTIC" ]
   sed -i "s:^SIM_TYPE.*:SIM_TYPE=3:" run_this_example_kernel.sh
 fi
 
+# save Par_file state
+if [ -e DATA/Par_file ]; then
+  cp -v DATA/Par_file DATA/Par_file.bak
+fi
+
 # runs simulation
-if [ "$do_kernel_script" == "1" ]; then
-  # kernel test script
-  ./run_this_example_kernel.sh
+if [ "${RUN_KERNEL}" == "true" ]; then
+  # use kernel script
+  ./run_this_example_kernel.sh | tee output.log
 else
   # default script
   ./run_this_example.sh
 fi
-
 # checks exit code
 if [[ $? -ne 0 ]]; then exit 1; fi
 
@@ -90,16 +114,30 @@ echo `date`
 echo
 
 # seismogram comparison
-if [ "${DEBUG}" == "true" ] || [ "$do_kernel_script" == "1" ]; then
-  # no comparisons
-  :       # do nothing
-else
-  # tests seismograms
+RUN_COMPARE=true
+# turn off for non-default runs
+if [ "${TESTCOV}" == "true" ]; then RUN_COMPARE=false; fi
+if [ "${DEBUG}" == "true" ]; then RUN_COMPARE=false; fi
+if [ "${RUN_KERNEL}" == "true" ]; then RUN_COMPARE=false; fi
+
+if [ "${RUN_COMPARE}" == "true" ]; then
   my_test
+else
+  # no comparisons
+  :     # do nothing
+fi
+# checks exit code
+if [[ $? -ne 0 ]]; then exit 1; fi
+
+# restore original Par_file
+if [ -e DATA/Par_file.bak ]; then
+  cp -v DATA/Par_file.bak DATA/Par_file
 fi
 
 # cleanup
 rm -rf OUTPUT_FILES*
+if [ -e DATABASES_MPI ]; then rm -rf DATABASES_MPI/; fi
+if [ -e SEM ]; then rm -rf SEM/; fi
 
 echo
 echo "all good"
